q2/requirements.txt
fastapi==0.116.0
uvicorn==0.35.0
sqlalchemy==2.0.41
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
pydantic==2.11.7
pydantic-settings==2.10.1
python-dotenv==1.1.1
email-validator==2.2.0
slowapi==0.1.9
redis==5.0.1
itsdangerous==2.1.2
bleach==6.1.0 